Refactor placeholders and support placeholders in dns records
